+++ src/main/kotlin/space/kscience/gradle/KScienceMPPlugin.kt
@@ -2,8 +2,14 @@ package space.kscience.gradle
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.kotlin.dsl.apply
+import org.gradle.kotlin.dsl.configure
 import org.gradle.kotlin.dsl.hasPlugin
+import org.gradle.kotlin.dsl.invoke
+import org.jetbrains.dokka.gradle.DokkaPlugin
+import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinMultiplatformPlugin
+import space.kscience.gradle.internal.applySettings
 
 public open class KScienceMPPlugin : Plugin<Project> {
     override fun apply(project: Project): Unit = project.run {
@@ -13,6 +19,36 @@ public open class KScienceMPPlugin : Plugin<Project> {
         } else {
             logger.info("Kotlin MPP plugin is already present")
         }
-        project.configureKScience()
+
+        registerKScienceExtension(::KScienceMppExtension)
+
+        configure<KotlinMultiplatformExtension> {
+
+            sourceSets {
+                getByName("commonMain"){
+                    dependencies {
+                        api(project.dependencies.platform("org.jetbrains.kotlin-wrappers:kotlin-wrappers-bom:${KScienceVersions.jsBom}"))
+                    }
+                }
+                getByName("commonTest"){
+                    dependencies {
+                        implementation(kotlin("test-common"))
+                        implementation(kotlin("test-annotations-common"))
+                    }
+                }
+            }
+
+            sourceSets.all {
+                languageSettings.applySettings()
+            }
+
+            if (explicitApi == null) explicitApiWarning()
+        }
+
+
+        // apply dokka for all projects
+        if (!plugins.hasPlugin("org.jetbrains.dokka")) {
+            apply<DokkaPlugin>()
+        }
     }
 }