+++ src/main/kotlin/space/kscience/gradle/KScienceVersions.kt
@@ -1,7 +1,7 @@
 package space.kscience.gradle
 
 
-import org.gradle.api.JavaVersion
+import org.gradle.jvm.toolchain.JavaLanguageVersion
 import org.tomlj.Toml
 
 /**
@@ -24,7 +24,7 @@ public object KScienceVersions {
     public val jsBom: String get() = toml.getString("versions.jsBom")!!
     internal val junit: String get() = toml.getString("versions.junit")!!
 
-    public val JVM_TARGET: JavaVersion = JavaVersion.VERSION_11
+    public val JVM_TARGET: JavaLanguageVersion = JavaLanguageVersion.of(11)
 
     public object Serialization {
         public val xmlVersion: String get() = toml.getString("versions.xmlutil")!!